Remove temporary Makefile
